+++ api/core/v2/user.go
@@ -58,3 +58,8 @@ func UserFields(r Resource) map[string]string {
 		"user.groups":   strings.Join(resource.Groups, ","),
 	}
 }
+
+// SetNamespace sets the namespace of the resource.
+func (u *User) SetNamespace(namespace string) {
+	return
+}