+++ CHANGELOG.md
@@ -25,6 +25,9 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 ### Changed
 - Eventd has been refactored. Users should not perceive any changes, but a
 substantial amount of business logic has been moved into other packages.
+- The `sensuctl create` command now accepts resources without a declared
+namespace. If the namespace is omitted, the resource will be created in the
+current namespace, or overridden by the `--namespace` flag.
 - Eventd now uses a constant number of requests to etcd when working with
 silenced entries, instead of a number that is proportional to the number of
 subscriptions in a check.